feat/services: routers to get, post, put

+++ src/services/services-service.ts
@@ -1,3 +1,4 @@
+import { notFound } from "../errors/errors";
 import serviceRepository from "../repositories/service-repository";
 
 async function createService(name: string, duration: number, price: number) {
@@ -19,12 +20,18 @@ async function listServices() {
 }
 
 async function updateService(id: number, name: string, duration: number, price: number) {
+  const serviceExistent = await serviceRepository.findById(id);
+  if(!serviceExistent) throw notFound();
+
   const service = await serviceRepository.updateService(id, name, duration, price);
 
   return service;
 }
 
 async function deleteService(id: number) {
+  const serviceExistent = await serviceRepository.findById(id);
+  if(!serviceExistent) throw notFound();
+
   const service = await serviceRepository.deleteService(id);
 
   return service;
